Improve compatibility with newer versions of CGI
HTML generation apparently used to rely on hash ordering: the attribute
name 'class' was used. However, if this showed up as the first parameter
in the argument list for a CGI HTML function, it could be mistaken as
e.g. a name attribute for an input element. Add a dash everywhere to
ensure the attribute name can never be mistaken to be some other kind of
parameter.